requirements.txt
requests~=2.31.0
loguru~=0.7.2
python-dotenv~=1.0.1
confluent-kafka~=2.5.3
tgcrypto~=1.2.5
telethon~=1.37.0